updated ai dev new template for openai-chat-streaming-js (now can use either openai or azure)
