+++ src/ai/.x/templates/openai-chat-streaming-js/package.json
@@ -1,15 +1,14 @@
 {
-    "name": "openai-chat-streaming",
-    "version": "1.0.0",
-    "description": "",
-    "main": "Main.js",
-    "scripts": {
-      "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "author": "",
-    "license": "MIT",
-    "dependencies": {
-      "@azure/openai": "1.0.0-beta.10"
-    }
+  "name": "openai-chat-streaming",
+  "version": "1.0.0",
+  "description": "",
+  "main": "Main.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "openai": "^4.0.0"
   }
-  
+}
